# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: tea121-lite backend
  version: 0.1.0
paths:
...
  /api/v1/detectors:
    get:
      tags:
      - catalog
      summary: Detectors
      operationId: detectors_api_v1_detectors_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_DetectorDescriptor_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/rule-packs:
    get:
      tags:
      - catalog
      summary: Rule Packs
      operationId: rule_packs_api_v1_rule_packs_get
      parameters:
      - name: detector_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Detector Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_RulePackDescriptor_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/projects/{project_id}/evaluations:
    post:
      tags:
      - evaluations
      summary: Create Evaluation
      operationId: create_evaluation_api_v1_projects__project_id__evaluations_post
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvaluationCreate'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluationOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - evaluations
      summary: List Evaluations
      operationId: list_evaluations_api_v1_projects__project_id__evaluations_get
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_EvaluationOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/evaluations/{evaluation_id}:
    get:
      tags:
      - evaluations
      summary: Get Evaluation
      operationId: get_evaluation_api_v1_evaluations__evaluation_id__get
      parameters:
      - name: evaluation_id
        in: path
        required: true
        schema:
          type: string
          title: Evaluation Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluationOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/evaluations/{evaluation_id}/cancel:
    post:
      tags:
      - evaluations
      summary: Cancel Evaluation
      operationId: cancel_evaluation_api_v1_evaluations__evaluation_id__cancel_post
      parameters:
      - name: evaluation_id
        in: path
        required: true
        schema:
          type: string
          title: Evaluation Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EvaluationOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/evaluations/{evaluation_id}/cases:
    get:
      tags:
      - evaluations
      summary: Cases
      operationId: cases_api_v1_evaluations__evaluation_id__cases_get
      parameters:
      - name: evaluation_id
        in: path
        required: true
        schema:
          type: string
          title: Evaluation Id
      - name: detector_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Detector Id
      - name: rule_pack_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Rule Pack Id
      - name: cwe_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cwe Id
      - name: family
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Family
      - name: violation_kind
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Violation Kind
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_EvaluationCaseOut_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/evaluations/{evaluation_id}/matrix:
    get:
      tags:
      - evaluations
      summary: Matrix
      operationId: matrix_api_v1_evaluations__evaluation_id__matrix_get
      parameters:
      - name: evaluation_id
        in: path
        required: true
        schema:
          type: string
          title: Evaluation Id
      - name: cwe_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Cwe Id
      - name: family
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Family
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatrixOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    DetectorDescriptor:
      properties:
        id:
          type: string
          title: Id
        version:
          type: string
          title: Version
        name:
          type: string
          title: Name
        supported_cwes:
          items:
            type: string
          type: array
          title: Supported Cwes
        enabled:
          type: boolean
          title: Enabled
      type: object
      required:
      - id
      - version
      - name
      - supported_cwes
      - enabled
      title: DetectorDescriptor
...
    EvaluationCaseOut:
      properties:
        id:
          type: string
          title: Id
        evaluation_id:
          type: string
          title: Evaluation Id
        case_name:
          type: string
          title: Case Name
        cwe_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Cwe Id
        family:
          anyOf:
          - type: string
          - type: 'null'
          title: Family
        bad_run_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Bad Run Id
        good_run_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Good Run Id
        bad_outcome:
          type: string
          title: Bad Outcome
        good_outcome:
          type: string
          title: Good Outcome
        classification:
          anyOf:
          - type: string
          - type: 'null'
          title: Classification
      type: object
      required:
      - id
      - evaluation_id
      - case_name
      - cwe_id
      - family
      - bad_run_id
      - good_run_id
      - bad_outcome
      - good_outcome
      - classification
      title: EvaluationCaseOut
    EvaluationCreate:
      properties:
        dataset_name:
          type: string
          minLength: 1
          title: Dataset Name
        dataset_version:
          type: string
          title: Dataset Version
          default: '1'
        detector_id:
          type: string
          title: Detector Id
        rule_pack_id:
          type: string
          title: Rule Pack Id
        filter:
          additionalProperties: true
          type: object
          title: Filter
        cases:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Cases
        manifest_path:
          anyOf:
          - type: string
          - type: 'null'
          title: Manifest Path
        max_cases:
          anyOf:
          - type: integer
            maximum: 10000.0
            minimum: 1.0
          - type: 'null'
          title: Max Cases
      type: object
      required:
      - dataset_name
      - detector_id
      - rule_pack_id
      title: EvaluationCreate
    EvaluationOut:
      properties:
        id:
          type: string
          title: Id
        project_id:
          type: string
          title: Project Id
        dataset_name:
          type: string
          title: Dataset Name
        dataset_version:
          type: string
          title: Dataset Version
        detector_id:
          type: string
          title: Detector Id
        detector_version:
          type: string
          title: Detector Version
        rule_pack_id:
          type: string
          title: Rule Pack Id
        rule_pack_version:
          type: string
          title: Rule Pack Version
        status:
          type: string
          title: Status
        filter:
          additionalProperties: true
          type: object
          title: Filter
        summary:
          additionalProperties: true
          type: object
          title: Summary
        created_at:
          title: Created At
        started_at:
          title: Started At
        finished_at:
          title: Finished At
      type: object
      required:
      - id
      - project_id
      - dataset_name
      - dataset_version
      - detector_id
      - detector_version
      - rule_pack_id
      - rule_pack_version
      - status
      - filter
      - summary
      - created_at
      title: EvaluationOut
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    MatrixOut:
      properties:
        detector_id:
          type: string
          title: Detector Id
        rule_pack_id:
          type: string
          title: Rule Pack Id
        cwe_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Cwe Id
        family:
          anyOf:
          - type: string
          - type: 'null'
          title: Family
        correct:
          type: integer
          title: Correct
        false_positive:
          type: integer
          title: False Positive
        false_negative:
          type: integer
          title: False Negative
        inverted:
          type: integer
          title: Inverted
        unsupported:
          type: integer
          title: Unsupported
        error:
          type: integer
          title: Error
        total:
          type: integer
          title: Total
        bad_recall:
          anyOf:
          - type: number
          - type: 'null'
          title: Bad Recall
        good_specificity:
          anyOf:
          - type: number
          - type: 'null'
          title: Good Specificity
        file_accuracy:
          anyOf:
          - type: number
          - type: 'null'
          title: File Accuracy
      type: object
      required:
      - detector_id
      - rule_pack_id
      - correct
      - false_positive
      - false_negative
      - inverted
      - unsupported
      - error
      - total
      title: MatrixOut
...
    Page_DetectorDescriptor_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/DetectorDescriptor'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
      type: object
      required:
      - items
      - total
      - limit
      - offset
      title: Page[DetectorDescriptor]
...
    Page_EvaluationCaseOut_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/EvaluationCaseOut'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
      type: object
      required:
      - items
      - total
      - limit
      - offset
      title: Page[EvaluationCaseOut]
    Page_EvaluationOut_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/EvaluationOut'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
      type: object
      required:
      - items
      - total
      - limit
      - offset
      title: Page[EvaluationOut]
...
    Page_RulePackDescriptor_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/RulePackDescriptor'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        limit:
          type: integer
          title: Limit
        offset:
          type: integer
          title: Offset
      type: object
      required:
      - items
      - total
      - limit
      - offset
      title: Page[RulePackDescriptor]
...
    RulePackDescriptor:
      properties:
        id:
          type: string
          title: Id
        detector_id:
          type: string
          title: Detector Id
        version:
          type: string
          title: Version
        name:
          type: string
          title: Name
        supported_cwes:
          items:
            type: string
          type: array
          title: Supported Cwes
        supported_families:
          items:
            type: string
          type: array
          title: Supported Families
        supported_violation_kinds:
          items:
            type: string
          type: array
          title: Supported Violation Kinds
      type: object
      required:
      - id
      - detector_id
      - version
      - name
      - supported_cwes
      - supported_families
      - supported_violation_kinds
      title: RulePackDescriptor
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError